+++ src/main/java/com/application/close/mqtt/service/MqttService.java
@@ -8,6 +8,8 @@ public interface MqttService {
 
 	MqttConnectResp connectWithAuth(int mqttParamId); // TCP + username/password
 
+	void disconnect(int mqttParamId);
+
 //	MqttConnectResp connectWithTLS(int mqttParamId); // TLS (SSL) connection
 //
 //	MqttConnectResp connectWithTLSAndAuth(int mqttParamId); // TLS + username/password