+++ src/main/java/com/application/close/bridge/BridgeService.java
@@ -1,0 +1,9 @@
+package com.application.close.bridge;
+
+public interface BridgeService {
+
+	void connectModToMqtt(int modTcpId, int mqttParamId);
+
+	void disconnectModFromMqtt(int modTcpId);
+
+}