+++ src/main/java/com/application/close/mqtt/controller/MqttParamController.java
@@ -6,7 +6,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -15,7 +14,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import com.application.close.mqtt.entity.MqttParam;
 import com.application.close.mqtt.payload.MqttParamPayload;
-import com.application.close.mqtt.payload.TopicPayload;
 import com.application.close.mqtt.service.MqttParamService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -67,17 +65,17 @@ public class MqttParamController {
 		return ResponseEntity.noContent().build();
 	}
 
-	@Operation(summary = "Add publish topics", description = "Adds or updates publish topics for a specific Mqtt param.")
-	@PatchMapping("/add/publishTopics")
-	public ResponseEntity<MqttParam> addPublishTopics(@Valid @RequestBody TopicPayload topicPayload) {
-		MqttParam updatedParam = paramService.addPublishTopics(topicPayload);
-		return ResponseEntity.ok(updatedParam);
-	}
-
-	@Operation(summary = "Add subscribe topics", description = "Adds or updates subscribe topics for a specific Mqtt param.")
-	@PatchMapping("/add/subscribeTopics")
-	public ResponseEntity<MqttParam> addSubscribeTopics(@Valid @RequestBody TopicPayload topicPayload) {
-		MqttParam updatedParam = paramService.addSubscribeTopics(topicPayload);
-		return ResponseEntity.ok(updatedParam);
-	}
+//	@Operation(summary = "Add publish topics", description = "Adds or updates publish topics for a specific Mqtt param.")
+//	@PatchMapping("/add/publishTopics")
+//	public ResponseEntity<MqttParam> addPublishTopics(@Valid @RequestBody TopicPayload topicPayload) {
+//		MqttParam updatedParam = paramService.addPublishTopics(topicPayload);
+//		return ResponseEntity.ok(updatedParam);
+//	}
+//
+//	@Operation(summary = "Add subscribe topics", description = "Adds or updates subscribe topics for a specific Mqtt param.")
+//	@PatchMapping("/add/subscribeTopics")
+//	public ResponseEntity<MqttParam> addSubscribeTopics(@Valid @RequestBody TopicPayload topicPayload) {
+//		MqttParam updatedParam = paramService.addSubscribeTopics(topicPayload);
+//		return ResponseEntity.ok(updatedParam);
+//	}
 }